# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Users & Tasks API",
    "description": "API to manage users and their tasks with proper relationships, filtering, and status tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Welcome message",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "operationId": "get_users_users_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string" },
                { "type": "null" }
              ],
              "description": "Filter users by name",
              "title": "Name"
            },
            "description": "Filter users by name"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response Get Users Users Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a user",
        "operationId": "create_user_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks",
        "operationId": "get_tasks_tasks_get",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string" },
                { "type": "null" }
              ],
              "description": "Filter by title",
              "title": "Title"
            },
            "description": "Filter by title"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string" },
                { "type": "null" }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "integer" },
                { "type": "null" }
              ],
              "description": "Filter by user ID",
              "title": "User Id"
            },
            "description": "Filter by user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskOut"
                  },
                  "title": "Response Get Tasks Tasks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a task",
        "operationId": "create_task_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreate": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Anirudh Raj"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "example": "anirudh@example.com"
          }
        }
      },
      "UserOut": {
        "type": "object",
        "required": ["name", "email", "id"],
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        }
      },
      "TaskCreate": {
        "type": "object",
        "required": ["title", "status", "user_id"],
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Build FastAPI app"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "example": "Set up endpoints for user/tasks"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "example": "pending"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "example": 1
          }
        }
      },
      "TaskOut": {
        "type": "object",
        "required": ["id", "title", "status", "user_id"],
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        }
      }
    }
  }
}